Remove query and insert statement parser / analyzer deprecated code
