--- fe/fe-core/src/main/java/com/starrocks/analysis/InsertSource.java
@@ -1,34 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package com.starrocks.analysis;
-
-// This struct make parser work simple
-// QueryStmt after insert and VALUES can hold in this Class
-// Now we only support query
-// Maybe can convert VALUES(v1, v2, v3) to "select v1, v2, v3" statement.
-public class InsertSource {
-    private QueryStmt queryStmt;
-
-    public InsertSource(QueryStmt queryStmt) {
-        this.queryStmt = queryStmt;
-    }
-
-    public QueryStmt getQueryStmt() {
-        return queryStmt;
-    }
-}
