+++ fe/fe-core/src/main/java/com/starrocks/analysis/Subquery.java
@@ -21,7 +21,6 @@
 
 package com.starrocks.analysis;
 
-import com.google.common.base.Preconditions;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.sql.analyzer.AST2SQL;
 import com.starrocks.sql.analyzer.SemanticException;
@@ -31,8 +30,6 @@ import com.starrocks.thrift.TExprNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.ArrayList;
-
 /**
  * Class representing a subquery. A Subquery consists of a QueryStmt and has
  * its own Analyzer context.
@@ -40,21 +37,11 @@ import java.util.ArrayList;
 public class Subquery extends Expr {
     private static final Logger LOG = LoggerFactory.getLogger(Subquery.class);
 
-    // The QueryStmt of the subquery.
-    protected QueryStmt stmt;
-    // A subquery has its own analysis context
-    protected Analyzer analyzer;
-
     // mark work way
     protected boolean useSemiAnti;
 
-    public Analyzer getAnalyzer() {
-        return analyzer;
-    }
+    private final QueryStatement queryStatement;
 
-    public QueryStmt getStatement() {
-        return stmt;
-    }
 
     public boolean isUseSemiAnti() {
         return useSemiAnti;
@@ -64,51 +51,13 @@ public class Subquery extends Expr {
         this.useSemiAnti = useSemiAnti;
     }
 
-    @Override
-    public String toSqlImpl() {
-        if (stmt != null) {
-            return "(" + stmt.toSql() + ")";
-        } else {
-            return "(" + AST2SQL.toString(queryStatement) + ")";
-        }
-    }
-
-    @Override
-    public String toDigestImpl() {
-        return "(" + stmt.toDigest() + ")";
-    }
-
-    /**
-     * C'tor that initializes a Subquery from a QueryStmt.
-     */
-    public Subquery(QueryStmt queryStmt) {
-        super();
-        Preconditions.checkNotNull(queryStmt);
-        stmt = queryStmt;
-        stmt.setNeedToSql(true);
-    }
-
-    /**
-     * Copy c'tor.
-     */
-    public Subquery(Subquery other) {
-        super(other);
-        stmt = other.stmt.clone();
-        // needToSql will affect the toSQL result, so we must clone it
-        stmt.setNeedToSql(other.stmt.needToSql);
-        analyzer = other.analyzer;
-    }
-
     public Subquery(QueryStatement queryStatement) {
         super();
         this.queryStatement = queryStatement;
     }
 
-    /**
-     * Analyzes the subquery in a child analyzer.
-     */
-    @Override
-    public void analyzeImpl(Analyzer parentAnalyzer) throws AnalysisException {
+    public QueryStatement getQueryStatement() {
+        return queryStatement;
     }
 
     @Override
@@ -116,42 +65,22 @@ public class Subquery extends Expr {
         return false;
     }
 
-    /**
-     * Check if the subquery's SelectStmt returns a single column of scalar type.
-     */
-    public boolean returnsScalarColumn() {
-        ArrayList<Expr> stmtResultExprs = stmt.getResultExprs();
-        if (stmtResultExprs.size() == 1 && stmtResultExprs.get(0).getType().isScalarType()) {
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * Returns true if the toSql() of the Subqueries is identical. May return false for
-     * equivalent statements even due to minor syntactic differences like parenthesis.
-     * TODO: Switch to a less restrictive implementation.
-     */
     @Override
     public boolean equals(Object o) {
         if (!super.equals(o)) {
             return false;
         }
 
-        if (queryStatement != null) {
-            if (((Subquery) o).getQueryStatement() == null) {
-                return false;
-            } else {
-                return o.equals(queryStatement);
-            }
+        if (((Subquery) o).getQueryStatement() == null) {
+            return false;
+        } else {
+            return o.equals(queryStatement);
         }
-
-        return stmt.toSql().equals(((Subquery) o).stmt.toSql());
     }
 
     @Override
     public Subquery clone() {
-        Subquery ret = new Subquery(this);
+        Subquery ret = new Subquery(queryStatement);
         LOG.debug("SUBQUERY clone old={} new={}",
                 System.identityHashCode(this),
                 System.identityHashCode(ret));
@@ -159,23 +88,21 @@ public class Subquery extends Expr {
     }
 
     @Override
+    protected void analyzeImpl(Analyzer analyzer) throws AnalysisException {
+
+    }
+
+    @Override
+    public String toSqlImpl() {
+        return "(" + AST2SQL.toString(queryStatement) + ")";
+    }
+
+    @Override
     protected void toThrift(TExprNode msg) {
     }
 
-    /**
-     * Below function is added by new analyzer
-     */
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context) throws SemanticException {
         return visitor.visitSubquery(this, context);
-    }
-
-    /**
-     * Analyzed subquery is store as QueryBlock
-     */
-    private QueryStatement queryStatement;
-
-    public QueryStatement getQueryStatement() {
-        return queryStatement;
     }
 }
