--- fe/fe-core/src/main/java/com/starrocks/analysis/WithClause.java
@@ -1,143 +1,0 @@
-// This file is made available under Elastic License 2.0.
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/fe/fe-core/src/main/java/org/apache/doris/analysis/WithClause.java
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package com.starrocks.analysis;
-
-import com.google.common.base.Joiner;
-import com.google.common.base.Preconditions;
-import com.google.common.collect.Lists;
-import com.starrocks.catalog.Database;
-import com.starrocks.catalog.View;
-import com.starrocks.common.AnalysisException;
-import com.starrocks.common.UserException;
-import com.starrocks.qe.ConnectContext;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-
-/**
- * Representation of the WITH clause that may appear before a query statement or insert
- * statement. A WITH clause contains a list of named view definitions that may be
- * referenced in the query statement that follows it.
- * <p>
- * Scoping rules:
- * A WITH-clause view is visible inside the query statement that it belongs to.
- * This includes inline views and nested WITH clauses inside the query statement.
- * <p>
- * Each WITH clause establishes a new analysis scope. A WITH-clause view definition
- * may refer to views from the same WITH-clause appearing to its left, and to all
- * WITH-clause views from outer scopes.
- * <p>
- * References to WITH-clause views are resolved inside out, i.e., a match is found by
- * first looking in the current scope and then in the enclosing scope(s).
- * <p>
- * Views defined within the same WITH-clause may not use the same alias.
- */
-public class WithClause implements ParseNode {
-    /////////////////////////////////////////
-    // BEGIN: Members that need to be reset()
-
-    private final ArrayList<View> views_;
-
-    // END: Members that need to be reset()
-    /////////////////////////////////////////
-
-    public WithClause(ArrayList<View> views) {
-        Preconditions.checkNotNull(views);
-        Preconditions.checkState(!views.isEmpty());
-        views_ = views;
-    }
-
-    /**
-     * Analyzes all views and registers them with the analyzer. Enforces scoping rules.
-     * All local views registered with the analyzer are have QueryStmts with resolved
-     * TableRefs to simplify the analysis of view references.
-     */
-    @Override
-    public void analyze(Analyzer analyzer) throws UserException {
-    }
-
-    /**
-     * C'tor for cloning.
-     */
-    private WithClause(WithClause other) {
-        Preconditions.checkNotNull(other);
-        views_ = Lists.newArrayList();
-        for (View view : other.views_) {
-            views_.add(new View(view.getName(), view.getQueryStmt().clone(),
-                    view.getOriginalColLabels()));
-        }
-    }
-
-    public void reset() {
-        for (View view : views_) {
-            view.getQueryStmt().reset();
-        }
-    }
-
-    public void getDbs(ConnectContext context, Map<String, Database> dbs) throws AnalysisException {
-        for (View view : views_) {
-            QueryStmt stmt = view.getQueryStmt();
-            stmt.getDbs(context, dbs);
-        }
-    }
-
-    @Override
-    public WithClause clone() {
-        return new WithClause(this);
-    }
-
-    @Override
-    public String toSql() {
-        List<String> viewStrings = Lists.newArrayList();
-        for (View view : views_) {
-            // Enclose the view alias and explicit labels in quotes if Hive cannot parse it
-            // without quotes. This is needed for view compatibility between Impala and Hive.
-            String aliasSql = ToSqlUtils.getIdentSql(view.getName());
-            if (view.hasColLabels()) {
-                aliasSql += "(" + Joiner.on(", ").join(
-                        ToSqlUtils.getIdentSqlList(view.getOriginalColLabels())) + ")";
-            }
-            viewStrings.add(aliasSql + " AS (" + view.getQueryStmt().toSql() + ")");
-        }
-        return "WITH " + Joiner.on(",").join(viewStrings);
-    }
-
-    public String toDigest() {
-        List<String> viewStrings = Lists.newArrayList();
-        for (View view : views_) {
-            // Enclose the view alias and explicit labels in quotes if Hive cannot parse it
-            // without quotes. This is needed for view compatibility between Impala and Hive.
-            String aliasSql = ToSqlUtils.getIdentSql(view.getName());
-            if (view.hasColLabels()) {
-                aliasSql += "(" + Joiner.on(", ").join(
-                        ToSqlUtils.getIdentSqlList(view.getOriginalColLabels())) + ")";
-            }
-            viewStrings.add(aliasSql + " as (" + view.getQueryStmt().toDigest() + ")");
-        }
-        return "with " + Joiner.on(",").join(viewStrings);
-    }
-
-    public List<View> getViews() {
-        return views_;
-    }
-}
