+++ fe/fe-core/src/main/java/com/starrocks/qe/StmtExecutor.java
@@ -33,11 +33,8 @@ import com.starrocks.analysis.DeleteStmt;
 import com.starrocks.analysis.DmlStmt;
 import com.starrocks.analysis.ExportStmt;
 import com.starrocks.analysis.Expr;
-import com.starrocks.analysis.InsertStmt;
 import com.starrocks.analysis.KillStmt;
-import com.starrocks.analysis.QueryStmt;
 import com.starrocks.analysis.RedirectStatus;
-import com.starrocks.analysis.SelectStmt;
 import com.starrocks.analysis.SetStmt;
 import com.starrocks.analysis.SetVar;
 import com.starrocks.analysis.ShowStmt;
@@ -103,6 +100,7 @@ import com.starrocks.sql.ast.CreateTableAsSelectStmt;
 import com.starrocks.sql.ast.DropHistogramStmt;
 import com.starrocks.sql.ast.DropStatsStmt;
 import com.starrocks.sql.ast.ExecuteAsStmt;
+import com.starrocks.sql.ast.InsertStmt;
 import com.starrocks.sql.ast.QueryStatement;
 import com.starrocks.sql.ast.SelectRelation;
 import com.starrocks.sql.ast.UseCatalogStmt;
@@ -248,8 +246,7 @@ public class StmtExecutor {
         }
 
         // this is a query stmt, but this non-master FE can not read, forward it to master
-        if ((parsedStmt instanceof QueryStmt || parsedStmt instanceof QueryStatement) &&
-                !GlobalStateMgr.getCurrentState().isLeader()
+        if (parsedStmt instanceof QueryStatement && !GlobalStateMgr.getCurrentState().isLeader()
                 && !GlobalStateMgr.getCurrentState().canRead()) {
             return true;
         }
@@ -281,10 +278,6 @@ public class StmtExecutor {
         }
     }
 
-    public boolean isQueryStmt() {
-        return parsedStmt != null && (parsedStmt instanceof QueryStmt || parsedStmt instanceof QueryStatement);
-    }
-
     public StatementBase getParsedStmt() {
         return parsedStmt;
     }
@@ -308,10 +301,7 @@ public class StmtExecutor {
             // support select hint e.g. select /*+ SET_VAR(query_timeout=1) */ sleep(3);
             if (parsedStmt != null) {
                 Map<String, String> optHints = null;
-                if (parsedStmt instanceof SelectStmt) {
-                    SelectStmt selectStmt = (SelectStmt) parsedStmt;
-                    optHints = selectStmt.getSelectList().getOptHints();
-                } else if (parsedStmt instanceof QueryStatement &&
+                if (parsedStmt instanceof QueryStatement &&
                         ((QueryStatement) parsedStmt).getQueryRelation() instanceof SelectRelation) {
                     SelectRelation selectRelation = (SelectRelation) ((QueryStatement) parsedStmt).getQueryRelation();
                     optHints = selectRelation.getSelectList().getOptHints();
@@ -382,7 +372,7 @@ public class StmtExecutor {
                 LOG.debug("no need to transfer to Leader. stmt: {}", context.getStmtId());
             }
 
-            if (parsedStmt instanceof QueryStmt || parsedStmt instanceof QueryStatement) {
+            if (parsedStmt instanceof QueryStatement) {
                 context.getState().setIsQuery(true);
 
                 // sql's blacklist is enabled through enable_sql_blacklist.
@@ -510,30 +500,8 @@ public class StmtExecutor {
                     // If this sql is in blacklist, show message.
                     SqlBlackList.verifying(originSql);
                 }
-
-                InsertStmt insertStmt = (InsertStmt) parsedStmt;
-                // The transaction of an insert operation begin at analyze phase.
-                // So we should abort the transaction at this finally block if it encounters exception.
-                if (insertStmt.isTransactionBegin() && context.getState().getStateType() == MysqlStateType.ERR) {
-                    try {
-                        String errMsg = Strings.emptyToNull(context.getState().getErrorMessage());
-                        if (insertStmt.getTargetTable() instanceof ExternalOlapTable) {
-                            ExternalOlapTable externalTable = (ExternalOlapTable) (insertStmt.getTargetTable());
-                            GlobalStateMgr.getCurrentGlobalTransactionMgr().abortRemoteTransaction(
-                                    externalTable.getSourceTableDbId(), insertStmt.getTransactionId(),
-                                    externalTable.getSourceTableHost(),
-                                    externalTable.getSourceTablePort(),
-                                    errMsg == null ? "unknown reason" : errMsg);
-                        } else {
-                            GlobalStateMgr.getCurrentGlobalTransactionMgr().abortTransaction(
-                                    insertStmt.getDbObj().getId(), insertStmt.getTransactionId(),
-                                    (errMsg == null ? "unknown reason" : errMsg));
-                        }
-                    } catch (Exception abortTxnException) {
-                        LOG.warn("errors when abort txn", abortTxnException);
-                    }
-                }
             }
+
             context.setSessionVariable(sessionVariableBackup);
         }
     }
@@ -632,21 +600,15 @@ public class StmtExecutor {
             }
         }
 
-        if (parsedStmt instanceof QueryStmt
-                || parsedStmt instanceof QueryStatement
-                || parsedStmt instanceof DmlStmt
-                || parsedStmt instanceof CreateTableAsSelectStmt) {
-            Preconditions.checkState(false, "Shouldn't reach here");
-        } else {
-            try {
-                parsedStmt.analyze(new Analyzer(context.getGlobalStateMgr(), context));
-            } catch (AnalysisException e) {
-                throw e;
-            } catch (Exception e) {
-                LOG.warn("Analyze failed because ", e);
-                throw new AnalysisException("Unexpected exception: " + e.getMessage());
-            }
+        try {
+            parsedStmt.analyze(new Analyzer(context.getGlobalStateMgr(), context));
+        } catch (AnalysisException e) {
+            throw e;
+        } catch (Exception e) {
+            LOG.warn("Analyze failed because ", e);
+            throw new AnalysisException("Unexpected exception: " + e.getMessage());
         }
+
     }
 
     // Because this is called by other thread
@@ -742,9 +704,7 @@ public class StmtExecutor {
         RowBatch batch;
         MysqlChannel channel = context.getMysqlChannel();
         boolean isOutfileQuery = false;
-        if (queryStmt instanceof QueryStmt) {
-            isOutfileQuery = ((QueryStmt) queryStmt).hasOutFileClause();
-        } else if (queryStmt instanceof QueryStatement) {
+        if (queryStmt instanceof QueryStatement) {
             isOutfileQuery = ((QueryStatement) queryStmt).hasOutFileClause();
         }
         boolean isSendFields = false;
