+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/IntersectRelation.java
@@ -1,8 +1,6 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 package com.starrocks.sql.ast;
 
-import com.starrocks.sql.optimizer.base.SetQualifier;
-
 import java.util.List;
 
 public class IntersectRelation extends SetOperationRelation {
