+++ fe/fe-core/src/test/java/com/starrocks/load/InsertOverwriteJobTest.java
@@ -3,7 +3,7 @@
 package com.starrocks.load;
 
 import com.google.common.collect.Lists;
-import com.starrocks.analysis.InsertStmt;
+import com.starrocks.sql.ast.InsertStmt;
 import mockit.Expectations;
 import mockit.Injectable;
 import org.junit.Assert;
