+++ fe/fe-core/src/test/java/com/starrocks/sql/analyzer/AdminShowTest.java
@@ -1,8 +1,8 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 package com.starrocks.sql.analyzer;
 
-import com.starrocks.analysis.InsertStmt;
 import com.starrocks.sql.ast.DescribeStmt;
+import com.starrocks.sql.ast.InsertStmt;
 import com.starrocks.utframe.UtFrameUtils;
 import org.junit.Assert;
 import org.junit.BeforeClass;
